% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newEuropeReviewR-package.R
\docType{package}
\name{newEuropeReviewR-package}
\alias{newEuropeReviewR}
\alias{newEuropeReviewR-package}
\title{newEuropeReviewR: Data package contains a complete run of the foreign policy weekly 'The New Europe'}
\description{
This R data package contains text files for the complete run of the foreign policy weekly _The New Europe_ that was published between 1916 and 1920.
}
\author{
\strong{Maintainer}: Stanislav Pejša \email{spejsa@purdue.edu} (\href{https://orcid.org/0000-0001-5057-262X}{ORCID})

}
\keyword{internal}
