% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tne_entities.R
\docType{data}
\name{tne_entities}
\alias{tne_entities}
\title{Data of nominal brown shrimp prices by size}
\format{
A tibble with 212,559 rows and 14 variables:
\describe{
\item{doc_id}{dbl Year price was recorded}
\item{sid}{dbl Month price was recorded. Ranges from 1-12 for January - December}
\item{tid}{dbl denoting monthly price for the 0-15/lb. size class}
\item{tid_end}{dbl denoting monthly price for the 15-20/lb. size class}
\item{entity_type}{dbl denoting monthly price for the 20-25/lb. size class}
\item{entity}{dbl denoting monthly price for the 25-30/lb. size class}
\item{author_norm}{dbl denoting monthly price for the 30-40/lb. size class}
\item{date}{dbl denoting monthly price of pieces of shrimp totalling a pound}
\item{type}{dbl denoting monthly price for the 25-30/lb. size class}
\item{volume}{dbl denoting monthly price for the 30-40/lb. size class}
\item{number}{dbl denoting monthly price for the 40-50/lb. size class}
\item{year}{dbl denoting monthly price for the 50-67/lb. size class}
\item{article_title}{dbl denoting monthly price of pieces of shrimp totalling a pound}
\item{taerget_country}{dbl denoting monthly price of pieces of shrimp totalling a pound}
}
}
\source{
Extracted, edited, and transformed into a dataset by Stanislav Pejša.
}
\usage{
tne_entities
}
\description{
Shrimp are classified by size, 0-15 shrimp per pound, 15-20 shrimp per pound, etc. A smaller number per pound indicates larger shrimp. Nominal prices are total monthly value of brown shrimp andings within size class divided by total monthly landings within the size class.
}
\examples{
  tne_entities
}
\keyword{datasets}
