% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newEuropeReviewR-package.R
\docType{package}
\name{newEuropeReviewR-package}
\alias{newEuropeReviewR}
\alias{newEuropeReviewR-package}
\title{newEuropeReviewR: The New Europe Data Package contains the complete run of the foreign policy weekly _The New Europe_}
\description{
This R data package contains text corpus of a complete run of the foreign policy weekly _The New Europe_ that was published between 1916 and 1921. The dataset allows for textual analysis of the weekly.
}
\seealso{
Useful links:
\itemize{
  \item \url{"https://github.com/standap/peaceBooksR"}
}

}
\author{
\strong{Maintainer}: Stanislav Pejša \email{spejsa@purdue.edu} (\href{https://orcid.org/0000-0001-5057-262X}{ORCID})

}
\keyword{internal}
